refactor(ext): extract backend seams — db, settings-store, task-dispatch, task-events, asset-token, plugin-registry
Mechanical moves of existing implementations into src/ext-default/ so a
shell can substitute serverless backends via src/ext/* (the established
@ext alias fallback). No behavior change for the open-source build:

- @ext/db: better-sqlite3 per-scope connection map moves out of
  src/db/index.ts (now a re-export + schema).
- @ext/settings-store: settings.json fs read/write; env-store keeps
  parsing/cleaning + the codec hook.
- @ext/task-dispatch: local in-process dispatch behind a dynamic import
  so remote dispatchers don't bundle the child_process chain.
- @ext/task-events: the in-process EventEmitter backend; emitter.ts
  keeps the TaskEvent type and re-exports the API (5 consumers
  unchanged).
- @ext/asset-token: the in-process token map, now async-signature so
  stateless (signed-token) backends fit; call sites await.
- @ext/plugin-registry: the disk scanner + dev watcher backend;
  plugins-registry.server.ts re-exports (7 consumers unchanged).

Verified: lint/typecheck/build pass; dev smoke over every seam
(settings roundtrip, workspace list on existing data, upload/serve,
engine-assets 401, registry scan) is behavior-identical.
