+++ src/app/api/task/wait/route.ts
@@ -1,10 +1,10 @@
+import { dispatchTask } from "@ext/task-dispatch";
 import type { NextRequest } from "next/server";
 import { isTerminalStatus } from "@/constants/task-status";
 import { jsonStringifyForSse } from "@/lib/json-sse";
 import { logger } from "@/lib/logger";
 import { getScope, runWithScope } from "@/lib/runtime/scope.server";
 import { isTaskRunning, onTaskEvent, type TaskEvent } from "@/lib/task/emitter";
-import { dispatchTask } from "@/lib/task/runner";
 
 /**
  * GET /api/task/wait?taskId=xxx&reconnect=false
